+++ tests/templates/test_subroutines/test_trotter.py
@@ -23,19 +23,9 @@ import pytest
 import pennylane as qml
 from pennylane import numpy as qnp
 from pennylane.math import allclose, get_interface
-from pennylane.templates.subroutines.trotter import (
-    _scalar,
-    _simplify,
-    _comm_error,
-    _spectral_norm,
-    _one_norm_error,
-    _flatten_trotter,
-    _recursive_flatten,
-    _compute_repetitions,
-    _generate_combinations,
-    _recursive_decomposition,
-    _recursive_nested_commutator,
-)
+from pennylane.resource.error import SpectralNormError
+from pennylane.templates.subroutines.trotter import _scalar, _recursive_decomposition
+
 
 test_hamiltonians = (
     qml.dot([1, 1, 1], [qml.PauliX(0), qml.PauliY(0), qml.PauliZ(1)]),
@@ -488,296 +478,62 @@ class TestPrivateFunctions:
             qml.equal(op1, op2) for op1, op2 in zip(decomp, expected_expansion)
         )  # Expected expression
 
-    @pytest.mark.parametrize(
-        "order, n, expected_val",
-        (
-            (1, 1, 1),
-            (1, 3, 3),
-            (2, 3, 6),
-            (4, 1, 10),
-            (4, 10, 100),  # Computed by hand
-        ),
-    )
-    def test_compute_repetitions(self, order, n, expected_val):
-        assert _compute_repetitions(order, n) == expected_val
-
-    @pytest.mark.parametrize(
-        "op, fast, expected_norm",
-        (
-            (qml.s_prod(0, qml.I(0)), False, 0),
-            (qml.s_prod(1.23, qml.X(0)), False, 1.23),
-            (
-                qml.sum(qml.Z(1), qml.s_prod(-1.23, qml.X(0))),
-                True,
-                2.23,
-            ),
-            (
-                qml.sum(qml.Z(1), qml.s_prod(-1.23, qml.I(0))),
-                False,
-                2.23,
-            ),
-        ),
-    )
-    def test_spectral_norm_pauli(self, op, fast, expected_norm):
-        assert _spectral_norm(op, fast=fast) == expected_norm
-
-    @pytest.mark.parametrize(
-        "op, fast, expected_norm",  # computed by hand
-        (
-            (qml.Hadamard(0), False, 1),
-            (qml.Hadamard(0), True, 1),
-            (qml.RX(1.23, 0), True, 1),
-            (qml.s_prod(-0.5, qml.Hadamard(0)), False, 0.5),
-        ),
-    )
-    def test_spectral_norm_non_pauli(self, op, fast, expected_norm):
-        if not fast:
-            assert _spectral_norm(op, False) == expected_norm
-        else:
-            assert _spectral_norm(op, True) >= expected_norm
-
-    combination_data = (
-        (0, 0, ()),
-        (0, 123, ()),
-        (1, 0, ((0,),)),
-        (5, 0, ((0, 0, 0, 0, 0),)),
-        (1, 123, ((123,),)),
-        (2, 1, ((0, 1), (1, 0))),
-        (2, 2, ((0, 2), (1, 1), (2, 0))),
-        (2, 3, ((0, 3), (1, 2), (2, 1), (3, 0))),
-        (3, 1, ((0, 0, 1), (0, 1, 0), (1, 0, 0))),
-    )
-
-    @pytest.mark.parametrize("num_var, req_sum, expected_tup", combination_data)
-    def test_generate_combinations(self, num_var, req_sum, expected_tup):
-        assert _generate_combinations(num_var, req_sum) == expected_tup
-
-    @pytest.mark.parametrize(
-        "A, B, alpha, final_op",  # computed by hand
-        (
-            (qml.X(0), qml.Y(0), 0, qml.Y(0)),
-            (qml.X(0), qml.Y(0), 1, qml.s_prod(2j, qml.Z(0))),
-            (qml.X(0), qml.Y(0), 2, qml.s_prod(4, qml.Y(0))),
-            (qml.X(0), qml.Y(0), 3, qml.s_prod(8j, qml.Z(0))),
-            (qml.X(0), qml.Y(1), 3, qml.s_prod(0, qml.I(wires=(0, 1)))),
-            (
-                qml.RX(1.23, 0),
-                qml.RZ(-0.5, 0),
-                1,
-                qml.sum(
-                    qml.prod(qml.RX(1.23, 0), qml.RZ(-0.5, 0)),
-                    qml.s_prod(-1, qml.prod(qml.RZ(-0.5, 0), qml.RX(1.23, 0))),
-                ),
-            ),
-            (
-                qml.s_prod(-0.5, qml.prod(qml.X(0), qml.RX(123, 1))),
-                qml.RZ(-0.5, 2),
-                4,
-                qml.s_prod(0, qml.I(wires=(0, 1, 2))),
-            ),
-        ),
-    )
-    def test_recursive_nested_commutator(self, A, B, alpha, final_op):
-        m_expected = qml.matrix(final_op)
-        computed_op = _recursive_nested_commutator(A, B, alpha)
-
-        try:
-            m_computed = qml.matrix(computed_op, wire_order=final_op.wires)
-        except qml.operation.DecompositionUndefinedError:
-            pr = computed_op.pauli_rep
-            pr.simplify()
-            m_computed = pr.to_mat(wire_order=final_op.wires.tolist())
-
-        assert qnp.allclose(m_computed, m_expected)
-
-    @pytest.mark.parametrize(
-        "order, num_ops, t, expected_indicies_and_coeffs",
-        (
-            (1, 2, 1, ([0, 1], [1, 1])),
-            (1, 3, 0.25, ([0, 1, 2], [0.25, 0.25, 0.25])),
-            (2, 3, 1, ([0, 1, 2, 2, 1, 0], [0.5, 0.5, 0.5, 0.5, 0.5, 0.5])),
-            (2, 2, 8, ([0, 1, 1, 0], [4, 4, 4, 4])),
-            (
-                4,
-                2,
-                1,
-                (
-                    [0, 1, 1, 0] * 5,
-                    [0.5 * p_4] * 8 + [0.5 * (1 - 4 * p_4)] * 4 + [0.5 * p_4] * 8,
-                ),
-            ),
-        ),
-    )
-    def test_recursive_flatten(self, order, num_ops, t, expected_indicies_and_coeffs):
-        computed_indicies, computed_coeffs = _recursive_flatten(order, num_ops, t)
-        expected_indicies, expected_coeffs = expected_indicies_and_coeffs
-
-        assert computed_coeffs == expected_coeffs
-        assert computed_indicies == expected_indicies
-
-    @pytest.mark.parametrize(
-        "index_lst, coeffs_lst, simplified_index_lst, simplified_coeffs_lst",
-        (  # computed by hand
-            (
-                [0, 1, 2, 3],
-                [123, -0.45, -6.7, 89],
-                [0, 1, 2, 3],
-                [123, -0.45, -6.7, 89],
-            ),  # no simplification
-            ([0, 1, 1, 0], [1.23, -4, 5, 6.7], [0, 1, 0], [1.23, 1, 6.7]),
-            (
-                [0, 1, 1, 0, 0, 1, 1, 0],
-                [1, 2, 3, 4, -5, -6, -7, -8],
-                [0, 1, 0, 1, 0],
-                [1, 5, -1, -13, -8],
-            ),
-        ),
-    )
-    def test_simplify_flatten(
-        self, index_lst, coeffs_lst, simplified_index_lst, simplified_coeffs_lst
-    ):
-        computed_index_lst, computed_coeffs_lst = _simplify(index_lst, coeffs_lst)
-        assert computed_index_lst == simplified_index_lst
-        assert computed_coeffs_lst == simplified_coeffs_lst
-
-    @pytest.mark.parametrize(
-        "order, num_ops, n, expected_indicies_and_coeffs",
-        (
-            (1, 3, 1, ([0, 1, 2], [1] * 3)),
-            (1, 3, 3, ([0, 1, 2, 0, 1, 2, 0, 1, 2], [1 / 3] * 9)),
-            (2, 3, 1, ([0, 1, 2, 1, 0], [0.5, 0.5, 1, 0.5, 0.5])),
-            (
-                2,
-                2,
-                3,
-                (
-                    [0, 1, 0, 1, 0, 1, 0],
-                    [1 / 6, 2 / 6, 2 / 6, 2 / 6, 2 / 6, 2 / 6, 1 / 6],
-                ),
-            ),
-            (
-                4,
-                2,
-                2,
-                (
-                    [0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0],
-                    [
-                        0.25 * p_4,
-                        p_4 / 2,
-                        p_4 / 2,
-                        p_4 / 2,
-                        0.25 * (1 - 3 * p_4),
-                        (1 - 4 * p_4) / 2,
-                        0.25 * (1 - 3 * p_4),
-                        p_4 / 2,
-                        p_4 / 2,
-                        p_4 / 2,
-                        p_4 / 2,
-                        p_4 / 2,
-                        p_4 / 2,
-                        p_4 / 2,
-                        0.25 * (1 - 3 * p_4),
-                        (1 - 4 * p_4) / 2,
-                        0.25 * (1 - 3 * p_4),
-                        p_4 / 2,
-                        p_4 / 2,
-                        p_4 / 2,
-                        0.25 * p_4,
-                    ],
-                ),
-            ),
-        ),
-    )
-    def test_flatten_trotter(self, order, num_ops, n, expected_indicies_and_coeffs):
-        computed_indicies, computed_coeffs = _flatten_trotter(num_ops, order, n)
-        expected_indicies, expected_coeffs = expected_indicies_and_coeffs
-
-        assert qnp.allclose(computed_coeffs, expected_coeffs)
-        assert computed_indicies == expected_indicies
-
 
 class TestError:
-    """Test the error estimation functionality"""
+    """Test the error method of the TrotterProduct class"""
 
-    one_norm_error_dict = {
-        (1, 1): lambda a1, a2, t: (t * (abs(a1) + abs(a2))) ** 2,
-        (2, 1): lambda a1, a2, t: (5 / 4) * (t * (abs(a1) + abs(a2))) ** 2,
-        (1, 2): lambda a1, a2, t: (3 / 2) * (t * (abs(a1) + abs(a2))) ** 3,
-        (2, 2): lambda a1, a2, t: (65 / 24) * (t * (abs(a1) + abs(a2))) ** 3,
-        (1, 4): lambda a1, a2, t: ((10**5 + 1) / 120) * (t * (abs(a1) + abs(a2))) ** 5,
-    }
+    @pytest.mark.parametrize("fast", (True, False))
+    def test_invalid_method(self, fast):
+        """Test that passing an invalid method raises an error."""
+        method = "crazy"
+        op = qml.TrotterProduct(qml.sum(qml.X(0), qml.Y(0)), 1.23)
+        with pytest.raises(ValueError, match=f"The '{method}' method is not supported"):
+            _ = op.error(method, fast=fast)
 
+    def test_one_norm_error_method(self):
+        """Test that the one-norm error method works as expected."""
+        op = qml.TrotterProduct(qml.sum(qml.X(0), qml.Y(0)), time=0.05, order=4)
+        expected_error = ((10**5 + 1) / 120) * (0.1**5)
+
+        for computed_error in (
+            op.error(method="one-norm"),
+            op.error(method="one-norm", fast=False),
+        ):
+            assert isinstance(computed_error, SpectralNormError)
+            assert qnp.isclose(computed_error.error, expected_error)
+
+    def test_commutator_error_method(self):
+        """Test that the commutator error method works as expected."""
+        op = qml.TrotterProduct(qml.sum(qml.X(0), qml.Y(0)), time=0.05, order=2)
+        expected_error = (32 / 3) * (0.05**3)
+
+        for computed_error in (
+            op.error(method="commutator"),
+            op.error(method="commutator", fast=False),
+        ):
+            assert isinstance(computed_error, SpectralNormError)
+            assert qnp.isclose(computed_error.error, expected_error)
+
+    @pytest.mark.all_interfaces
     @pytest.mark.parametrize(
-        "steps, order",
-        (
-            (1, 1),
-            (2, 1),
-            (1, 2),
-            (2, 2),
-            (1, 4),
-        ),
+        "method, expected_error", (("one-norm", 0.009140625), ("commutator", 8e-3))
     )
-    @pytest.mark.parametrize(
-        "h_coeffs",
-        (
-            (1, 1),
-            (0.5, 2),
-            (1.23, -0.45),
-            (1j, 3.14j),
-        ),
-    )
-    @pytest.mark.parametrize("time", (1, 0.5, 0.25, 0.01))
-    def test_one_norm_error(self, steps, order, time, h_coeffs):
-        h_ops = (qml.s_prod(h_coeffs[0], qml.X(0)), qml.s_prod(h_coeffs[1], qml.Z(0)))
-        expected_error = self.one_norm_error_dict[(steps, order)](h_coeffs[0], h_coeffs[1], time)
+    @pytest.mark.parametrize("interface", ("autograd", "jax", "tensorflow", "torch"))
+    def test_error_interfaces(self, method, interface, expected_error):
+        """Test that the error method works with all interfaces"""
 
-        computed_error = _one_norm_error(h_ops, time, order, steps, fast=False)
-        assert qnp.isclose(computed_error, expected_error, atol=1e-12)
+        time = qml.math.array(0.1, like=interface)
+        coeffs = qml.math.array([1.0, 0.5], like=interface)
 
-        computed_error_fast = _one_norm_error(h_ops, time, order, steps, fast=True)
-        assert (
-            qnp.isclose(computed_error_fast, expected_error, atol=1e-12)
-            or computed_error_fast > expected_error
-        )
+        hamiltonian = qml.dot(coeffs, [qml.X(0), qml.Y(0)])
 
-    commutator_error_dict = {
-        (1, 1): lambda a1, a2, t: 4 * t**2 * abs(a1) * abs(a2),
-        (2, 1): lambda a1, a2, t: 8 * t**2 * abs(a1) * abs(a2),
-        (1, 2): lambda a1, a2, t: (16 / 3) * t**3 * abs(a1) * abs(a2) * (abs(a1) + abs(a2)),
-        (2, 2): lambda a1, a2, t: (32 / 3) * t**3 * abs(a1) * abs(a2) * (abs(a1) + abs(a2)),
-    }
+        op = qml.TrotterProduct(hamiltonian, time, n=2, order=2)
+        computed_error = op.error(method=method)
 
-    @pytest.mark.parametrize(
-        "steps, order",
-        (
-            (1, 1),
-            (2, 1),
-            (1, 2),
-            (2, 2),
-        ),
-    )
-    @pytest.mark.parametrize(
-        "h_coeffs",
-        (
-            (1, 1),
-            (0.5, 2),
-            (1.23, -0.45),
-            (1j, 3.14j),
-        ),
-    )
-    @pytest.mark.parametrize("time", (1, 0.5, 0.25, 0.01))
-    def test_comm_error(self, steps, order, time, h_coeffs):
-        h_ops = (qml.s_prod(h_coeffs[0], qml.X(0)), qml.s_prod(h_coeffs[1], qml.Z(0)))
-        expected_error = self.commutator_error_dict[(steps, order)](h_coeffs[0], h_coeffs[1], time)
+        assert isinstance(computed_error, SpectralNormError)
+        assert qml.math.get_interface(computed_error.error) == interface
 
-        computed_error = _comm_error(h_ops, time, order, steps, fast=False)
-        assert qnp.isclose(computed_error, expected_error, atol=1e-12)
-
-        computed_error_fast = _comm_error(h_ops, time, order, steps, fast=True)
-        assert (
-            qnp.isclose(computed_error_fast, expected_error, atol=1e-12)
-            or computed_error_fast > expected_error
-        )
+        assert qnp.isclose(computed_error.error, qml.math.array(expected_error, like=interface))
 
 
 class TestDecomposition:
